rename post route for posts
